Change: Make audit report component naming more consistent
Use consistent names for the audit report components as it is difficult
to understand the purpose of a component form its name.
