+++ src/web/pages/reports/AuditReportDetailsPage.jsx
@@ -16,14 +16,16 @@ import {isActive} from 'gmp/models/task';
 import {first} from 'gmp/utils/array';
 import {isDefined, hasValue} from 'gmp/utils/identity';
 import withDownload from 'web/components/form/withDownload';
+import PageTitle from 'web/components/layout/PageTitle';
 import Reload, {
   NO_RELOAD,
   USE_DEFAULT_RELOAD_INTERVAL_ACTIVE,
 } from 'web/components/loading/Reload';
 import withDialogNotification from 'web/components/notification/withDialogNotification';
+import FilterProvider from 'web/entities/FilterProvider';
 import useGmp from 'web/hooks/useGmp';
 import useTranslation from 'web/hooks/useTranslation';
-import Page from 'web/pages/reports/DeltaDetailsContent';
+import Page from 'web/pages/reports/AuditReportDetailsContent';
 import DownloadReportDialog from 'web/pages/reports/DownloadReportDialog';
 import ReportDetailsFilterDialog from 'web/pages/reports/ReportDetailsFilterDialog';
 import TargetComponent from 'web/pages/targets/TargetComponent';
@@ -31,8 +33,8 @@ import {
   loadAllEntities as loadFilters,
   selector as filterSelector,
 } from 'web/store/entities/filters';
-import {loadDeltaAuditReport} from 'web/store/entities/report/actions';
-import {deltaAuditReportSelector} from 'web/store/entities/report/selectors';
+import {loadAuditReportWithThreshold} from 'web/store/entities/report/actions';
+import {auditReportSelector} from 'web/store/entities/report/selectors';
 import {
   loadAllEntities as loadReportConfigs,
   selector as reportConfigsSelector,
@@ -41,6 +43,8 @@ import {
   loadAllEntities as loadReportFormats,
   selector as reportFormatsSelector,
 } from 'web/store/entities/reportformats';
+import {pageFilter as setPageFilter} from 'web/store/pages/actions';
+import getPage from 'web/store/pages/selectors';
 import {
   loadReportComposerDefaults,
   saveReportComposerDefaults,
@@ -52,17 +56,24 @@ import {
   getReportComposerDefaults,
   getUsername,
 } from 'web/store/usersettings/selectors';
+import {create_pem_certificate} from 'web/utils/Cert';
 import compose from 'web/utils/Compose';
 import PropTypes from 'web/utils/PropTypes';
 import {generateFilename} from 'web/utils/Render';
 
-const log = logger.getLogger('web.pages.report.deltadetailspage');
+const log = logger.getLogger('web.pages.auditreport./DetailsPage');
 
 const DEFAULT_FILTER = Filter.fromString(
-  'levels=hmlg rows=100 min_qod=70 first=1 sort-reverse=compliant',
+  'levels=hmlg rows=100 min_qod=70 first=1 compliance_levels=yniu sort=compliant',
 );
 
+export const AUDIT_REPORT_RESET_FILTER = RESET_FILTER.copy()
+  .setSortOrder('sort')
+  .setSortBy('compliant');
+
 const REPORT_FORMATS_FILTER = Filter.fromString('active=1 and trust=1 rows=-1');
+
+const getReportPageName = id => `report-${id}`;
 
 const getTarget = (entity = {}) => {
   const {report = {}} = entity;
@@ -70,20 +81,26 @@ const getTarget = (entity = {}) => {
   return task.target;
 };
 
-const DeltaAuditReportDetails = props => {
+const AuditReportDetailsPage = props => {
   const [showFilterDialog, setShowFilterDialog] = useState(false);
   const [showDownloadReportDialog, setShowDownloadReportDialog] =
     useState(false);
-  const [reportFormatId, setReportFormatId] = useState();
-  const [isUpdating, setIsUpdating] = useState(false);
-  // storeAsDefault is set in SaveDialogContent
-  // eslint-disable-next-line no-unused-vars
-  const [storeAsDefault, setStoreAsDefault] = useState();
-
   const [sorting, setSorting] = useState({
     results: {
       sortField: 'compliant',
       sortReverse: true,
+    },
+    hosts: {
+      sortField: 'compliant',
+      sortReverse: true,
+    },
+    os: {
+      sortField: 'compliant',
+      sortReverse: true,
+    },
+    tlscerts: {
+      sortField: 'dn',
+      sortReverse: false,
     },
     errors: {
       sortField: 'error',
@@ -91,22 +108,37 @@ const DeltaAuditReportDetails = props => {
     },
   });
 
+  const [entity, setEntity] = useState();
+  const [resultsCounts, setResultsCounts] = useState();
+  const [hostsCounts, setHostsCounts] = useState();
+  const [operatingSystemsCounts, setOperatingSystemsCounts] = useState();
+  const [tlsCertificatesCounts, setTlsCertificatesCounts] = useState();
+  const [reportFormatId, setReportFormatId] = useState();
+  const [errorsCounts, setErrorsCounts] = useState();
+  const [reportFilter, setReportFilter] = useState();
+  const [isUpdating, setIsUpdating] = useState(false);
+  // storeAsDefault is set in SaveDialogContent
+  // eslint-disable-next-line no-unused-vars
+  const [storeAsDefault, setStoreAsDefault] = useState();
+
   const [_] = useTranslation();
   const gmp = useGmp();
   const dispatch = useDispatch();
-  const params = useParams();
-  const {id: reportId, deltaid: deltaReportId} = params;
+  const params = useParams('/audit/:id');
+  const {id: reportId} = params;
 
-  const reportFormatsSel = useSelector(reportFormatsSelector);
-  const reportConfigsSel = useSelector(reportConfigsSelector);
-  const reportFormats = reportFormatsSel?.getAllEntities(REPORT_FORMATS_FILTER);
-  const reportConfigs = reportConfigsSel?.getAllEntities(ALL_FILTER);
-  const userDefaultFilterSel = useSelector(
-    rootState => getUserSettingsDefaultFilter(rootState, 'result'),
-    shallowEqual,
-  );
-  const resultDefaultFilter = userDefaultFilterSel?.getFilter('result');
-  const reportComposerDefaults = useSelector(getReportComposerDefaults);
+  const pSelector = useSelector(getPage, shallowEqual);
+  const pageFilter = pSelector?.getFilter(getReportPageName(reportId));
+
+  const [selectedEntity, reportError, isLoading] = useSelector(state => {
+    const reportSel = auditReportSelector(state);
+    return [
+      reportSel?.getEntity(reportId, pageFilter),
+      reportSel?.getEntityError(reportId, pageFilter),
+      reportSel?.isLoadingEntity(reportId, pageFilter),
+    ];
+  }, shallowEqual);
+
   const userDefaultsSelector = useSelector(
     getUserSettingsDefaults,
     shallowEqual,
@@ -114,17 +146,18 @@ const DeltaAuditReportDetails = props => {
   const reportExportFileName = userDefaultsSelector?.getValueByName(
     'reportexportfilename',
   );
+
+  const reportFormatsSel = useSelector(reportFormatsSelector);
+  const reportConfigsSel = useSelector(reportConfigsSelector);
+  const reportFormats = reportFormatsSel?.getAllEntities(REPORT_FORMATS_FILTER);
+  const reportConfigs = reportConfigsSel?.getAllEntities(ALL_FILTER);
+  const reportComposerDefaults = useSelector(getReportComposerDefaults);
+  const userDefaultFilterSel = useSelector(
+    rootState => getUserSettingsDefaultFilter(rootState, 'result'),
+    shallowEqual,
+  );
+  const resultDefaultFilter = userDefaultFilterSel?.getFilter();
   const username = useSelector(getUsername);
-  const filterSel = useSelector(filterSelector);
-  const filters = filterSel?.getAllEntities(RESULTS_FILTER_FILTER);
-  const [entity, entityError] = useSelector(state => {
-    const deltaSel = deltaAuditReportSelector(state);
-    return [
-      deltaSel?.getEntity(reportId, deltaReportId),
-      deltaSel?.getError(reportId, deltaReportId),
-    ];
-  });
-  const isLoading = !isDefined(entity);
 
   useEffect(() => {
     dispatch(loadUserSettingDefaults(gmp)());
@@ -132,8 +165,31 @@ const DeltaAuditReportDetails = props => {
     dispatch(loadReportFormats(gmp)(REPORT_FORMATS_FILTER));
     dispatch(loadReportConfigs(gmp)(ALL_FILTER));
     dispatch(loadReportComposerDefaults(gmp)());
+
+    if (isDefined(selectedEntity)) {
+      setEntity(entity);
+      updateReportCounts(selectedEntity);
+      setReportFilter(props.reportFilter);
+      setIsUpdating(false);
+    } else {
+      setIsUpdating(true);
+    }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  useEffect(() => {
+    if (isDefined(selectedEntity)) {
+      // update only if a new report is available to avoid having no report
+      // when the filter changes
+      setEntity(selectedEntity);
+      updateReportCounts(selectedEntity);
+      setReportFilter(props.reportFilter);
+      setIsUpdating(false);
+    } else {
+      // report is not in the store and is currently loaded
+      setIsUpdating(true);
+    }
+  }, [selectedEntity, props.reportFilter]);
 
   useEffect(() => {
     if (
@@ -167,18 +223,45 @@ const DeltaAuditReportDetails = props => {
   useEffect(() => {
     load();
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [reportId, deltaReportId]);
+  }, [reportId]);
 
-  const load = filter => {
+  const updateReportCounts = reportEntity => {
+    const {report = {}} = reportEntity;
+    const {
+      results = {},
+      hosts = {},
+      operatingSystems = {},
+      tlsCertificates = {},
+      errors = {},
+    } = report;
+
+    if (isDefined(results.counts)) {
+      setResultsCounts(results.counts);
+    }
+    if (isDefined(hosts.counts)) {
+      setHostsCounts(hosts.counts);
+    }
+    if (isDefined(operatingSystems.counts)) {
+      setOperatingSystemsCounts(operatingSystems.counts);
+    }
+    if (isDefined(tlsCertificates.counts)) {
+      setTlsCertificatesCounts(tlsCertificates.counts);
+    }
+    if (isDefined(errors.counts)) {
+      setErrorsCounts(errors.counts);
+    }
+  };
+
+  const load = newFilter => {
     log.debug('Loading report', {
-      filter,
+      newFilter,
     });
-    const {reportFilter: prevFilter} = props;
+    const {reportFilter: filter} = props;
 
-    setIsUpdating(!isDefined(prevFilter) || !prevFilter.equals(filter));
+    setIsUpdating(!isDefined(filter) || !filter.equals(newFilter)); // show update indicator if filter has changed
 
     props
-      .reload(filter)
+      .reload(newFilter)
       .then(() => {
         setIsUpdating(false);
       })
@@ -207,7 +290,7 @@ const DeltaAuditReportDetails = props => {
   };
 
   const handleFilterRemoveClick = () => {
-    handleFilterChange(RESET_FILTER);
+    handleFilterChange(AUDIT_REPORT_RESET_FILTER);
   };
 
   const handleFilterResetClick = () => {
@@ -221,7 +304,7 @@ const DeltaAuditReportDetails = props => {
   const handleAddToAssets = () => {
     const {showSuccessMessage, reportFilter: filter} = props;
 
-    gmp.auditreport.addAssets(entity, {filter}).then(() => {
+    gmp.auditreport.addAssets(selectedEntity, {filter}).then(() => {
       showSuccessMessage(
         _(
           'Report content added to Assets with QoD>=70% and Overrides enabled.',
@@ -234,7 +317,7 @@ const DeltaAuditReportDetails = props => {
   const handleRemoveFromAssets = () => {
     const {showSuccessMessage, reportFilter: filter} = props;
 
-    gmp.auditreport.removeAssets(entity, {filter}).then(() => {
+    gmp.auditreport.removeAssets(selectedEntity, {filter}).then(() => {
       showSuccessMessage(_('Report content removed from Assets.'));
       reload();
     }, handleError);
@@ -257,12 +340,18 @@ const DeltaAuditReportDetails = props => {
   };
 
   const handleReportDownload = state => {
-    const {reportFilter, onDownload} = props;
+    const {reportFilter: filter, onDownload} = props;
 
-    const {includeNotes, includeOverrides, reportFormatId, storeAsDefault} =
-      state;
+    const {
+      includeNotes,
+      includeOverrides,
 
-    const newFilter = reportFilter.copy();
+      reportFormatId,
+
+      storeAsDefault,
+    } = state;
+
+    const newFilter = filter.copy();
     newFilter.set('notes', includeNotes);
     newFilter.set('overrides', includeOverrides);
 
@@ -285,22 +374,21 @@ const DeltaAuditReportDetails = props => {
       : 'unknown'; // unknown should never happen but we should be save here
 
     return gmp.auditreport
-      .download(entity, {
+      .download(selectedEntity, {
         reportFormatId,
-        deltaReportId,
         filter: newFilter,
       })
       .then(response => {
         setShowDownloadReportDialog(false);
         const {data} = response;
         const filename = generateFilename({
-          creationTime: entity.creationTime,
+          creationTime: selectedEntity.creationTime,
           extension,
           fileNameFormat: reportExportFileName,
-          id: entity.id,
-          modificationTime: entity.modificationTime,
+          id: selectedEntity.id,
+          modificationTime: selectedEntity.modificationTime,
           reportFormat: report_format?.name,
-          resourceName: entity.task.name,
+          resourceName: selectedEntity.task.name,
           resourceType: 'report',
           username,
         });
@@ -309,16 +397,28 @@ const DeltaAuditReportDetails = props => {
       }, handleError);
   };
 
+  const handleTlsCertificateDownload = cert => {
+    const {onDownload} = props;
+
+    const {data, serial} = cert;
+
+    onDownload({
+      filename: 'tls-cert-' + serial + '.pem',
+      mimetype: 'application/x-x509-ca-cert',
+      data: create_pem_certificate(data),
+    });
+  };
+
   const handleFilterCreated = filter => {
     load(filter);
     dispatch(loadFilters(gmp)(RESULTS_FILTER_FILTER));
   };
 
   const handleFilterDecreaseMinQoD = () => {
-    const {reportFilter} = props;
+    const {reportFilter: filter} = props;
 
-    if (reportFilter.has('min_qod')) {
-      const lfilter = reportFilter.copy();
+    if (filter.has('min_qod')) {
+      const lfilter = filter.copy();
       lfilter.set('min_qod', 30);
       load(lfilter);
     }
@@ -330,43 +430,65 @@ const DeltaAuditReportDetails = props => {
     const sortReverse =
       sortField === prev.sortField ? !prev.sortReverse : false;
 
-    const newSorting = {
+    const newSort = {
       ...sorting,
       [name]: {
         sortField,
         sortReverse,
       },
     };
-    setSorting(newSorting);
+    setSorting(newSort);
   };
 
   const loadTarget = () => {
-    const target = getTarget(entity);
+    const target = getTarget(selectedEntity);
     return gmp.target.get({id: target.id});
   };
 
-  const {reportFilter, showError, showErrorMessage, showSuccessMessage} = props;
+  const {showError, showErrorMessage, showSuccessMessage} = props;
 
-  const {report} = entity || {};
+  const report = isDefined(entity) ? entity.report : undefined;
+
+  const threshold = gmp.settings.reportResultsThreshold;
+
+  const showThresholdMessage =
+    isDefined(report) && report.results.counts.filtered > threshold;
+
+  const [filters, isLoadingFilters] = useSelector(state => {
+    const filterSel = filterSelector(state);
+    return [
+      filterSel?.getAllEntities(RESULTS_FILTER_FILTER),
+      filterSel?.isLoadingAllEntities(RESULTS_FILTER_FILTER),
+    ];
+  });
 
   return (
     <React.Fragment>
+      <PageTitle title={_('Report Details')} />
       <TargetComponent onError={handleError}>
         {({edit}) => (
           <Page
             audit={true}
             entity={entity}
-            entityError={entityError}
-            filter={reportFilter}
+            errorsCounts={errorsCounts}
             filters={filters ? filters : []}
+            hostsCounts={hostsCounts}
             isLoading={isLoading}
+            isLoadingFilters={isLoadingFilters}
             isUpdating={isUpdating}
+            operatingSystemsCounts={operatingSystemsCounts}
+            pageFilter={pageFilter}
+            reportError={reportError}
+            reportFilter={reportFilter}
             reportId={reportId}
+            resetFilter={AUDIT_REPORT_RESET_FILTER}
+            resultsCounts={resultsCounts}
             showError={showError}
             showErrorMessage={showErrorMessage}
             showSuccessMessage={showSuccessMessage}
             sorting={sorting}
             task={isDefined(report) ? report.task : undefined}
+            tlsCertificatesCounts={tlsCertificatesCounts}
             onAddToAssetsClick={handleAddToAssets}
             onError={handleError}
             onFilterChanged={handleFilterChange}
@@ -382,6 +504,7 @@ const DeltaAuditReportDetails = props => {
             onTargetEditClick={() =>
               loadTarget().then(response => edit(response.data))
             }
+            onTlsCertificateDownloadClick={handleTlsCertificateDownload}
           />
         )}
       </TargetComponent>
@@ -389,7 +512,7 @@ const DeltaAuditReportDetails = props => {
         <ReportDetailsFilterDialog
           audit={true}
           createFilterType="result"
-          delta={true}
+          delta={false}
           filter={reportFilter}
           onClose={handleFilterDialogClose}
           onFilterChanged={handleFilterChange}
@@ -405,7 +528,9 @@ const DeltaAuditReportDetails = props => {
           includeOverrides={reportComposerDefaults.includeOverrides}
           reportConfigs={reportConfigs}
           reportFormats={reportFormats}
+          showThresholdMessage={showThresholdMessage}
           storeAsDefault={storeAsDefault}
+          threshold={threshold}
           onClose={handleCloseDownloadReportDialog}
           onSave={handleReportDownload}
         />
@@ -414,8 +539,7 @@ const DeltaAuditReportDetails = props => {
   );
 };
 
-DeltaAuditReportDetails.propTypes = {
-  defaultFilter: PropTypes.filter,
+AuditReportDetailsPage.propTypes = {
   location: PropTypes.object.isRequired,
   reload: PropTypes.func.isRequired,
   reportFilter: PropTypes.filter,
@@ -433,11 +557,25 @@ const reloadInterval = report =>
     : NO_RELOAD; // report doesn't change anymore. no need to reload
 
 const load =
-  ({defaultFilter, reportId, deltaReportId, dispatch, gmp, reportFilter}) =>
+  ({
+    defaultFilter,
+    reportId,
+
+    dispatch,
+    gmp,
+    params,
+    pageFilter,
+    reportFilter,
+  }) =>
   filter => {
     if (!hasValue(filter)) {
       // use loaded filter after initial loading
       filter = reportFilter;
+    }
+
+    if (!hasValue(filter)) {
+      // use filter from store
+      filter = pageFilter;
     }
 
     if (!hasValue(filter)) {
@@ -449,69 +587,69 @@ const load =
       // use fallback filter
       filter = DEFAULT_FILTER;
     }
-
-    // to avoid confusion of loaded results with different sort terms and
-    // directions, always load the report with sort=name from gvmd (the user's
-    // sort term will be handled by GSA in the browser)
-    filter.delete('sort-reverse');
-    filter.set('sort', 'name');
-    return dispatch(
-      loadDeltaAuditReport(gmp)(reportId, deltaReportId, filter),
-    ).then(() =>
-      dispatch(loadDeltaAuditReport(gmp)(reportId, deltaReportId, filter)),
-    );
+    dispatch(setPageFilter(getReportPageName(params.id), filter));
+    return dispatch(loadAuditReportWithThreshold(gmp)(reportId, {filter}));
   };
 
-const DeltaAuditReportDetailsWrapper = ({defaultFilter, ...props}) => {
-  const gmp = useGmp();
+const ReportDetailsWrapper = props => {
   const dispatch = useDispatch();
-  const params = useParams('/auditreport/delta/:id/:deltaid');
+  const gmp = useGmp();
+  const params = useParams();
 
-  const {id: reportId, deltaid: deltaReportId} = params;
-  const deltaSel = useSelector(deltaAuditReportSelector, shallowEqual);
-  const entity = deltaSel.getEntity(reportId, deltaReportId);
+  const {id: reportId} = params;
+  const reportSel = useSelector(auditReportSelector, shallowEqual);
+  const pSelector = useSelector(getPage, shallowEqual);
+
+  const pageFilter = pSelector.getFilter(getReportPageName(reportId));
+  const entity = reportSel.getEntity(reportId, pageFilter);
   const reportFilter = entity?.report?.filter;
 
   return (
-    <Reload
-      load={load({
-        ...props,
-        dispatch,
-        gmp,
-        defaultFilter,
-        reportFilter,
-        reportId,
-        deltaReportId,
-      })}
-      name="auditreport"
-      reload={load({
-        ...props,
-        dispatch,
-        gmp,
-        defaultFilter,
-        reportFilter,
-        reportId,
-        deltaReportId,
-      })}
-      reloadInterval={() => reloadInterval(entity)}
+    <FilterProvider
+      fallbackFilter={DEFAULT_FILTER}
+      gmpName="result"
+      pageName={`report-${reportId}`}
     >
-      {({reload}) => (
-        <DeltaAuditReportDetails
-          {...props}
-          defaultFilter={defaultFilter}
-          reload={reload}
-          reportFilter={reportFilter}
-        />
+      {({filter}) => (
+        <Reload
+          load={load({
+            ...props,
+            dispatch,
+            gmp,
+            params,
+            defaultFilter: filter,
+            reportFilter,
+            reportId,
+            pageFilter,
+          })}
+          name={`report-${reportId}`}
+          reload={load({
+            ...props,
+            dispatch,
+            gmp,
+            params,
+            defaultFilter: filter,
+            reportFilter,
+            reportId,
+            pageFilter,
+          })}
+          reloadInterval={() => reloadInterval(entity)}
+        >
+          {({reload}) => (
+            <AuditReportDetailsPage
+              {...props}
+              defaultFilter={filter}
+              reload={reload}
+              reportFilter={reportFilter}
+            />
+          )}
+        </Reload>
       )}
-    </Reload>
+    </FilterProvider>
   );
-};
-
-DeltaAuditReportDetailsWrapper.propTypes = {
-  defaultFilter: PropTypes.filter,
 };
 
 export default compose(
   withDialogNotification,
   withDownload,
-)(DeltaAuditReportDetailsWrapper);
+)(ReportDetailsWrapper);
